refact and fix other
