+++ includes/minishell.h
@@ -14,40 +14,14 @@
 # define MINISHELL_H
 
 # include "libft.h"
+# include "lexer.h"
+# include "parser.h"
 # include <stdio.h>
 # include <stdlib.h>
 # include <readline/readline.h>
 # include <readline/history.h>
 
 # define SUCCESS 0
-
-typedef enum e_type
-{
-	TYPE_GENERAL,
-	TYPE_PIPE,
-	TYPE_AMPERSAND,
-	TYPE_REDIRECT,
-	TYPE_EOF,
-	TYPE_NEWLINE,
-}	t_type;
-
-typedef enum e_node_kind
-{
-	NODE_ARGUMENT,
-	NODE_PIPE,
-	NODE_AMPERSAND,
-	NODE_REDIRECT,
-}	t_node_kind;
-
-typedef struct s_token
-{
-	t_type			type;
-	struct s_token	*head;
-	struct s_token	*cur;
-	struct s_token	*next;
-	const char		*str;
-	size_t			len;
-}	t_token;
 
 //typedef struct s_env
 //{
@@ -73,15 +47,6 @@ typedef struct s_token
 //	const char		*str;
 //	size_t			len;
 //}	t_node;
-
-typedef struct s_node
-{
-	t_node_kind		kind;
-	struct s_node	*left;
-	struct s_node	*right;
-	const char		*word;
-}	t_node;
-
 
 int		display_prompt(void);
 void	launch_executable(const char *exe_path);
